move Searchable into plotbot util for general usage
