+++ util/searchable.go
@@ -1,4 +1,4 @@
-package testutils
+package util
 
 import (
 	"fmt"
@@ -30,6 +30,36 @@ func (ps Searchable) ContainsAll(ss ...string) bool {
 func (ps Searchable) ContainsAny(ss ...string) bool {
 	for _, s := range ss {
 		if ps.Contains(s) {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (ps Searchable) Includes(s string) bool {
+	for _, p := range ps {
+		if p == s {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (ps Searchable) IncludesAll(ss ...string) bool {
+	for _, s := range ss {
+		if !ps.Includes(s) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func (ps Searchable) IncludesAny(ss ...string) bool {
+	for _, s := range ss {
+		if ps.Includes(s) {
 			return true
 		}
 	}
